Move GPU code into direct
I'll add another subdirectory to GPU for Mojo soon.
