+++ sky/shell/gpu/direct/ganesh_surface.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "sky/shell/gpu/ganesh_surface.h"
+#include "sky/shell/gpu/direct/ganesh_surface.h"
 
 #include "base/logging.h"
 #include "third_party/skia/include/gpu/GrContext.h"
