+++ sky/shell/gpu/direct/ganesh_surface.h
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef SKY_SHELL_GPU_GANESH_SURFACE_H_
-#define SKY_SHELL_GPU_GANESH_SURFACE_H_
+#ifndef SKY_SHELL_GPU_DIRECT_GANESH_SURFACE_H_
+#define SKY_SHELL_GPU_DIRECT_GANESH_SURFACE_H_
 
 #include "base/memory/scoped_ptr.h"
 #include "sky/engine/wtf/RefPtr.h"
-#include "sky/shell/gpu/ganesh_context.h"
+#include "sky/shell/gpu/direct/ganesh_context.h"
 #include "third_party/skia/include/core/SkSurface.h"
 #include "ui/gfx/geometry/size.h"
 
@@ -38,4 +38,4 @@ class GaneshSurface {
 }  // namespace shell
 }  // namespace sky
 
-#endif  // SKY_SHELL_GPU_GANESH_SURFACE_H_
+#endif  // SKY_SHELL_GPU_DIRECT_GANESH_SURFACE_H_
